rename _dispersion module
