+++ src/bmi/samplers/_matrix_utils.py
